requirements.txt

fastapi
uvicorn
python-dotenv
opencv-python
numpy
requests
Pillow
python-multipart